Ai_video_optimizer/requirements.txt
# --- Audio Transcription ---
# faster-whisper==0.10.0
torch>=1.13.1
torchvision>=0.13.0
numpy<2

# --- Scene Detection ---
scenedetect[opencv]==0.6.2
opencv-python==4.9.0.80
opencv-python-headless>=4.9.0.80  # Optional: use this if no GUI

# --- Visual Scoring (CLIP model) ---
git+https://github.com/openai/CLIP.git
Pillow>=9.0.0

# --- Video Processing ---
moviepy==1.0.3
ffmpeg-python==0.2.0

# --- OpenAI + Environment ---
openai>=1.3.0
python-dotenv>=1.0.0

# --- CLI & Utility ---
tqdm>=4.64.0

# --- Optional: For Groq API access ---
groq>=0.3.0

# --- Image editing ---
rembg #removes bg from logo

onnxruntime